requirements.txt
asgiref==3.10.0
beautifulsoup4==4.14.2
bootstrap4==0.1.0
contourpy==1.3.0
cycler==0.12.1
Django==4.2.25
django-bootstrap4==25.2
django-environ==0.12.0
fonttools==4.60.1
importlib_resources==6.5.2
kiwisolver==1.4.7
matplotlib==3.9.4
mysql-connector-python==9.4.0
mysqlclient==2.2.7
numpy==2.0.2
packaging==25.0
pillow==11.3.0
PyMySQL==1.1.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
typing_extensions==4.15.0
zipp==3.23.0
